Add support for the 'tags' clusterssh config file
ClusterSSH allows one to define clusters directly in the 'clusters' config file
or indirectly, with tags (that are cluster names) in the 'tags' config file.

This patch add support for the 'tags' file that was ignored.
It havily uses BASH extentions (mainly arrays), hence the changed shebang.
Excluded hosts can now also be clusters.
In this case, all hosts of these clusters will be excluded.